user props sent endpoint created, as well as props model to findPropsSent
