+++ data/model/props.js
@@ -10,6 +10,7 @@ module.exports = {
     update,
     updateKey,
     findByUserID,
+    findByPropsSent,
     findByDateRange
 }
 
@@ -20,6 +21,7 @@ function add(dbTable, obj) {
 function find() {
     return db(dbTable)
 }
+
 
 function findByDateRange(obj) {
     const { year, month, fk_from_workspace_profile_id } = obj
@@ -42,6 +44,11 @@ function findById(id) {
     return db(dbTable)
         .where({ id })
         .first()
+}
+
+function findByPropsSent(fk_from_workspace_profile_id) {
+    return db(dbTable)
+        .where({ fk_from_workspace_profile_id })
 }
 
 function remove(id) {
