+++ server/server.js
@@ -11,8 +11,7 @@ const oauth = require('./routes/oauth')
 const authZero = require('./auth/authZero')
 const authError = require('./auth/authError')
 const SPAAuth = require('../middleware/SPAauth')
-const userData = require('./routes/userData')
-const teamData = require('./routes/workspaceData')
+const propsData = require('./routes/propsData')
 
 server.use(express.json())
 server.use(helmet())
@@ -25,8 +24,7 @@ server.use('/slackActions', button)
 server.use('/info', info)
 server.use('/oauth', oauth)
 server.use('/protected', authZero)
-server.use('/user-data', userData)
-server.use('/team-data', teamData)
+server.use('/props-data', propsData)
 // Error handling for unauthorized users!
 server.use(authError)
